+++ advanced-project/app/Http/Controllers/CategoryController.php
@@ -58,6 +58,23 @@ class CategoryController extends Controller
     public function getById(Request $request, $id)
     {
         try {
+            // Check if the id is valid
+            if (!is_numeric($id)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Invalid Category ID'
+                ], 400);
+            }
+
+            // Check if Category exists
+            $category = Category::find($id);
+            if (!$category) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
             $category = Category::find($id);
             return response()->json([
                 'message' => $category,
@@ -70,21 +87,30 @@ class CategoryController extends Controller
     public function getByCategory(Request $request, $name)
     {
         try {
-            $report = Category::where('category', $name)->get();
-            if ($report->isEmpty()) {
+            // Check if the name is valid
+            if (!is_string($name) | !$name) {
                 return response()->json([
                     'success' => false,
-                    'message' => 'Report not found'
+                    'message' => 'Invalid Category name'
+                ], 400);
+            }
+
+            // Check if the category exists
+            $category = Category::where('category', $name)->get();
+            if ($category->isEmpty()) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Categories not found'
                 ], 404);
             } else {
                 return response()->json([
                     'success' => true,
-                    'data' => $report
+                    'data' => $category
                 ]);
             }
         } catch (\Exception $e) {
             return response()->json([
-                'message' => 'Error retrieving reports',
+                'message' => 'Error retrieving Categories',
                 'error' => $e->getMessage(),
             ], 500);
         }
@@ -93,6 +119,23 @@ class CategoryController extends Controller
     public function editById(Request $request, $id)
     {
         try {
+            // Check if the id is valid
+            if (!is_numeric($id) | !$id) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Invalid Category ID'
+                ], 400);
+            }
+
+            // Check if Category exists
+            $category = Category::find($id);
+            if (!$category) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
             $data = $request->only('category', 'category_description');
             $validator = Validator::make($data, [
                 'category' => 'required|string|min:3|max:27',
@@ -118,6 +161,23 @@ class CategoryController extends Controller
     public function editByName(Request $request, $name)
     {
         try {
+            // Check if the name is valid
+            if (!is_string($name)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Invalid Category Name'
+                ], 400);
+            }
+
+            // Check if Category exists
+            $category = Category::where('category',$name)->get();
+            if (!$category) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
             $data = $request->only('category', 'category_description');
             $validator = Validator::make($data, [
                 'category' => 'required|string|min:3|max:27',
@@ -143,6 +203,23 @@ class CategoryController extends Controller
     public function deleteCategoryById(Request $request, $id)
     {
         try {
+            // Check if the id is valid
+            if (!is_numeric($id) | !$id) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Invalid Category ID'
+                ], 400);
+            }
+
+            // Check if Category exists
+            $category = Category::find($id);
+            if (!$category) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
             $category = Category::find($id);
             $category->delete();
             return response()->json([
@@ -155,7 +232,24 @@ class CategoryController extends Controller
     public function deleteCategoryByName(Request $request, $name)
     {
         try {
-            $category = Category::where('category', $name);
+            // Check if the name is valid
+            if (!is_string($name)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Invalid Category Name'
+                ], 400);
+            }
+
+            // Check if Category exists
+            $category = Category::where('category',$name);
+            if (!$category) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
+            // $category = Category::where('category', $name);
             $category->delete();
             return response()->json([
                 'message' => 'Category deleted successfully',
